feat(.service)：apikey前后端调用

+++ llmops-common/src/main/java/com/emcikem/llm/common/entity/Paginator.java
@@ -15,7 +15,7 @@ public class Paginator {
 
     private Integer page_size;
 
-    private Integer total_page;
+    private Long total_page;
 
     private Integer total_record;
 }
